Hex viewer

+++ app/views.py
@@ -1,6 +1,7 @@
 import boto3
-import requests
-from io import BytesIO
+import urllib.request
+
+import glob
 from botocore.errorfactory import ClientError
 
 import os, re
@@ -25,19 +26,37 @@ def index(request):
         file_type = request.POST.get('file_type') if request.POST.get('file_type') != None else None
         
         if 'full_hex' in request.POST.keys():
-            current_url = ''
-            filename = request.session['current_url'] = current_url
-            relative_filename = filename.rsplit('/', 1)[-1]
-
-            s3 = boto3.client("s3")
             try:
-                # s3.head_object(Bucket='cysuite-bucket', Key='media/' + relative_filename)
-                s3.download_file('cysuite-bucket', 'media/' + relative_filename, 'media/payloads/' + relative_filename)
-                new_filename = 'media/payloads/' + relative_filename
-                response = requests.get(filename)
-                img = Image.open(BytesIO(response.content))
+                list_of_files = glob.glob('media/payloads/*')
+                if len(list_of_files) < 1:
+                    try:
+                        urllib.request.urlretrieve("https://cysuite-bucket.s3.us-west-2.amazonaws.com/media/default.png", "media/payloads/default.png")
+                    except:
+                        context['hex_dump'] = ''
+                        context['ipaddress'] = ip_address
+                        context['dimensions'] = '(0, 0)'
+                        context['file_type'] = 'None'
+                        context['file_size'] = '0 bytes'
+                        context['filename'] = 'None'
+                        context['extension'] = 'None'
+                        context['mime_type'] = 'None'
+                        context['byte_match'] = 'None'
+                        context['status'] = 'Not injected: Error downloading the image from S3 bucket'
+                        return render(request, 'index.html', context)
+                    else:
+                        relative_filename = 'default.png'
+                        latest_file = 'media/payloads/default.png'
+                        new_filename = 'media/payloads/default.png'
+                        context['status'] = 'Loaded default image'
+                else:
+                    latest_file = max(list_of_files, key=os.path.getctime)
+                    relative_filename = latest_file.rsplit('/', 1)[-1]
+                    new_filename = 'media/payloads/' + relative_filename
+                    context['status'] = 'Viewing full hex code'
+
+                img = Image.open(latest_file)
                 width, height = img.size
-                hex_dump = next(full_hex_viewer(filename))
+                hex_dump = next(full_hex_viewer(latest_file))
             
                 context['hex_dump'] = hex_dump
                 context['ipaddress'] = ip_address
@@ -48,24 +67,24 @@ def index(request):
                 context['extension'] = puremagic.magic_file(new_filename)[0].extension
                 context['mime_type'] = puremagic.magic_file(new_filename)[0].mime_type
                 context['byte_match'] = puremagic.magic_file(new_filename)[0].byte_match.decode('UTF-8','ignore').strip()
-                context['download'] = request.session['current_url']
-                context['status'] = 'Viewing full hex code'
                 context.update(csrf(request))
-                os.remove('media/payloads/' + relative_filename)
-                return render(request, 'dashboard/injector.html', context)
+                if new_filename != 'media/payloads/default.png':
+                    os.remove('media/payloads/' + relative_filename)
+                return render(request, 'index.html', context)
             except ClientError:
                 dimensions, filename = None, None
 
         elif (request.POST.get('filename') == None or request.POST.get('filename') == '' or '.' not in request.POST.get('filename')) and 'clear' not in request.POST.keys():
-            context['status'] = 'Not injected'
-            context['error_message'] = 'Filename does not follow the correct filename pattern'
+            context['status'] = 'Not injected: Filename does not follow the correct filename pattern'
             filename, dimensions = None, (None, None)
         elif (request.POST.get('filename') == None or request.POST.get('filename') == '' or '.' not in request.POST.get('filename')) and 'clear' in request.POST.keys():
             context['status'] = 'Cleared'
             filename, dimensions = None, (None, None)
         elif type(width) != int or type(height) != int or width < 0 or height < 0:
-            context['status'] = 'Not injected'
-            context['error_message'] = 'Dimension value should be an integer!'
+            context['status'] = 'Not injected: Dimension value should be an integer!'
+            filename, dimensions = None, (None, None)
+        elif file_type == 'PNG' and (width > 256 or height > 256):
+            context['status'] = 'Not injected: PNG files have a 256 maximum byte size!'
             filename, dimensions = None, (None, None)
         else:
             filename = 'media/payloads/' + request.POST.get('filename')
@@ -73,28 +92,26 @@ def index(request):
             filename, dimensions = injection.main()
 
         if filename != None and dimensions != None:
-            new_filename = re.sub(r'^.*?/', '', filename)
-            current_url = filename
+            list_files = glob.glob('media/payloads/*')
+            for file_ in list_files:
+                if filename != 'media/payloads/default.png' and file_ != filename:
+                    try:
+                        os.remove(file_)
+                    except:
+                        print("Tried to remove a non-existent payload image")
 
-            if new_filename != current_url and current_url != 'https://cysuite-bucket.s3.us-west-2.amazonaws.com/media/default.png':
-                try:
-                    del request.session['current_url']
-                except:
-                    print("Tried to remove a non-existent payload image")
-            request.session['current_url'] = current_url
-
-            final_filename = re.sub(r'^.*?/', '', new_filename)
+            relative_filename = filename.rsplit('/', 1)[-1]
             hex_dump = hex_viewer(filename)
             context['hex_dump'] = hex_dump
             context['ipaddress'] = ip_address
             context['dimensions'] = dimensions
             context['file_type'] = puremagic.magic_file(filename)[0].name 
             context['file_size'] = str(os.path.getsize(filename)) + ' bytes'
-            context['filename'] = final_filename
+            context['filename'] = relative_filename
             context['extension'] = puremagic.magic_file(filename)[0].extension
             context['mime_type'] = puremagic.magic_file(filename)[0].mime_type
             context['byte_match'] = puremagic.magic_file(filename)[0].byte_match.decode('UTF-8','ignore').strip()
-            context['download'] = request.session['current_url']
+            context['download'] = relative_filename
             context['status'] = 'Injected successfully'
             context['success_message'] = 'Your payload has been injected successfully!'
             return render(request, 'index.html', context)
